third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/debugger-set-breakpoint-regex-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests Debugger.setBreakpointByUrl with isRegex set to true.


Running: testSetNoneOfURLAndRegex
Either url or urlRegex or scriptHash must be specified.

Running: testSetBothURLAndRegex
Either url or urlRegex or scriptHash must be specified.

Running: testSetByRegex
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (debugger-set-breakpoint-regex.js:12)
Script execution resumed.
